src/src/Model/requirements.txt
pyyaml
pydantic
numpy
scipy
gpxpy
pyproj
math
pathlib
typing
matplotlib.pyplot